Add the release URL as an action ouput
This will allow other actions to compose workflows with that need access
to the URL of this release. In my case I wanted to get the  URL into a
slack message posted to the team when a release is published.

The output can be referenced by using the `steps.release.ouput.url` in
the workflow yaml:

    - name: Release
      id: release
      uses: softprops/action-gh-release@v1
      with:
        name: "My Release"
      env:
        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
    - name: Slack
      uses: csexton/slack-message-action@master
      with:
        message: New release posted at ${{ steps.release.outputs.url}}